requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
python-dotenv
bcrypt
python-jose
psycopg2-binary